refactor(conformance): consolidate shared AST infra and address PR review
Address review feedback on the P/O-series PR (cmgrote + sdk-review):

- Rename O001 StdlibJsonOverOrjson -> OrjsonOverStdlibJson (names the recommendation).
- Convert prescriptions/ and optimizations/ check modules to packages so each
  rule lives in its own file.
- Promote shared AST infrastructure (discover, directive parser, make_finding,
  make_cli_main, TOOL_VERSION) into a neutral _ast_common package and reverse the
  dependency: error_handling and actions_pinning now consume it instead of P/O
  reaching into E's private surface.
- Source the --tool-version default once from importlib.metadata; removes the
  per-CLI version drift.
- Collapse the four conformance CI jobs into one strategy.matrix; route the
  caller-controlled exclude-paths-* through quoted step-level env vars so they are
  never interpolated into the shell.
- Fix runner --exclude to match on path-component boundaries (tools excludes
  tools/ but not tools_extra/); add a regression test.
- Add registry-alignment invariants: doc-meta == rule series, checker series
  subset of rule series (L lags by design).
- Document the P-id/O-id non-migration policy (ids are permanent public contract).
